+++ src/CanvasRenderingContext2d.cc
@@ -828,8 +828,8 @@ NAN_METHOD(Context2d::PutImageData) {
   // clamp width at canvas size
   // Need to wrap std::min calls using parens to prevent macro expansion on
   // windows. See http://stackoverflow.com/questions/5004858/stdmin-gives-error
-  cols = (std::min)(sw, context->canvas()->getWidth() - dx);
-  rows = (std::min)(sh, context->canvas()->getHeight() - dy);
+  cols = (std::min)(sw, static_cast<int>(context->canvas()->getWidth()) - dx);
+  rows = (std::min)(sh, static_cast<int>(context->canvas()->getHeight()) - dy);
 
   if (cols <= 0 || rows <= 0) return;
